Rename ModelsTable.module.css to BrowseTable.module.css
