+++ frontend/src/metabase/browse/components/BrowseTable.module.css
